Modify example directory, modify code
Modify example directory to contain a supported bug that occurs in e-mails. Modify searchForHeader method to actually look for header a not only text.
Rename replaceDoubleNewLines method to replaceExtraNewLines.
Add indentation to match rest of the code.  Remove call to removeNewLines method that is not used and not valid in the code in this position that it was used. Add exception support when last value wasn't empty and the next line was continuation of the previous one.
Add replaceDoubleSpaces method.