Add optional progress bar, rename method
Added a progress bar. Slows down things so it's optional and disabledby default. Added some debug information like progress of the script. Renamed method readDirectory to processDirectory to better describe it's purpose.  Changed intendation for call to replaceDoubleSpaces so it's called once on a complete CSV line instead of being called every time in the loop.